FLEET-359 Fully plumbed in delete method
Also moved create out into its own file to avoid
mod.rs becoming huge
